enhance some business logic

+++ src/main/java/com/iot/admin/service/DeviceAdminServiceImpl.java
@@ -1,18 +1,28 @@
 package com.iot.admin.service;
 
 import com.iot.admin.dao.DeviceAdminRepository;
+import com.iot.dao.UserRepository;
 import com.iot.dto.SIMCardDto;
+import com.iot.exception.BusinessException;
 import com.iot.exception.DeviceNotFoundException;
-import com.iot.exception.UpdateDeviceStatusException;
-import com.iot.exception.WaitingForActivationDeviceNotFoundException;
+import com.iot.exception.UserNotFoundException;
+import com.iot.exception.UserPermissionException;
 import com.iot.mapper.SIMCardMapper;
 import com.iot.model.SIMCard;
+import com.iot.model.User;
 import com.iot.service.IotDeviceService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class DeviceAdminServiceImpl implements DeviceAdminService {
+
+    private static final String USER_ADMIN_ROLE = "Admin";
+
+    @Autowired
+    private UserRepository userRepository;
 
     @Autowired
     private IotDeviceService deviceService;
@@ -30,24 +40,36 @@ public class DeviceAdminServiceImpl implements DeviceAdminService {
     }
 
     @Override
-    public void removeDevice(Integer deviceId) {
+    public void removeDeviceByDeviceId(Integer deviceId, Integer userId) {
+        checkingUserPermissions(userId);
         adminRepository.deleteById(deviceId);
     }
 
     @Override
-    public SIMCardDto updateDeviceStatus(SIMCardDto simCardDto) throws UpdateDeviceStatusException {
-        SIMCard existingSimCard = adminRepository.findBySIMNumberAndOperationCode()
+    public SIMCardDto updateDeviceStatus(SIMCardDto simCardDto, Integer userId) throws BusinessException {
+
+        checkingUserPermissions(userId);
+
+        SIMCard existingSimCard = adminRepository.findBySIMNumberAndOperationCode(simCardDto.getSimNumber(), simCardDto.getOperation().getOperationCode())
                 .orElseThrow(() -> new DeviceNotFoundException());
 
-        SIMCard simToBeUpdated = new SIMCard();
-        simToBeUpdated.setStatus(simCardDto.getStatus());
-        simToBeUpdated.setCountryName(existingSimCard.getCountryName());
-        simToBeUpdated.setSimNumber(existingSimCard.getSimNumber());
-        simToBeUpdated.setOperation(existingSimCard.getOperation());
+        existingSimCard.setId(existingSimCard.getId());
+        existingSimCard.setStatus(simCardDto.getStatus());
+        existingSimCard.setCountryName(existingSimCard.getCountryName());
+        existingSimCard.setSimNumber(existingSimCard.getSimNumber());
+        existingSimCard.setOperation(existingSimCard.getOperation());
 
-        SIMCard updatedDeviceStatus = adminRepository.save(simToBeUpdated);
-        return simCardMapper.mapEntityToDTO(updatedDeviceStatus);
+        SIMCard updatedDevice = adminRepository.save(existingSimCard);
+        return simCardMapper.mapEntityToDTO(updatedDevice);
     }
 
+    private void checkingUserPermissions(Integer userId) {
 
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException());
+
+        if (!user.getRole().equalsIgnoreCase(USER_ADMIN_ROLE)) {
+            throw new UserPermissionException();
+        }
+    }
 }
